# docs/openapi/programmatic.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CredTrail Institution Integration API",
    "version": "1.0.0",
    "description": "Issue, revoke, track completion, discover templates, and reconcile institution credentials. Supply x-api-key. List pages sort by immutable ID ascending; pass nextCursor unchanged and retain the same filters. Dates are inclusive UTC calendar dates. Only email identities use case-insensitive matching. Authenticated responses are never cached."
  },
  "security": [
    {
      "IntegrationKey": []
    }
  ],
  "paths": {
    "/v1/programmatic/issue": {
      "post": {
        "operationId": "issueBadge",
        "summary": "Queue badge issuance",
        "x-required-scope": "queue.issue",
        "description": "idempotencyKey is required. Matching replays return the original operationId and assertionId. Reuse with different content returns 409. Poll statusUrl with operations.read; acceptance is not completion.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Persisted or replayed command. Follow statusUrl using your integration key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "Canonical status URL.",
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, invalid parameters, or unknown request fields (invalid_request).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, revoked, or invalid integration key (api_key_required, invalid_api_key).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Wrong institution, missing permission, invalid key scopes, or missing write owner.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found in this institution (operation_not_found, template_not_found, assertion_not_found).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict, archived template, or missing usable artwork (idempotency_conflict, template_archived, artwork_required).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Expiry must be later than the issue date (invalid_expiry).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Object storage is temporarily unavailable (storage_unavailable).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/programmatic/revoke": {
      "post": {
        "operationId": "revokeBadge",
        "summary": "Queue badge revocation",
        "x-required-scope": "queue.revoke",
        "description": "idempotencyKey is required. Matching replays return the original operationId and assertionId. Reuse with different content returns 409. Poll statusUrl with operations.read; acceptance is not completion.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Persisted or replayed command. Follow statusUrl using your integration key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "Canonical status URL.",
                "schema": {
                  "type": "string",
                  "format": "uri"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, invalid parameters, or unknown request fields (invalid_request).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, revoked, or invalid integration key (api_key_required, invalid_api_key).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Wrong institution, missing permission, invalid key scopes, or missing write owner.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found in this institution (operation_not_found, template_not_found, assertion_not_found).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict, archived template, or missing usable artwork (idempotency_conflict, template_archived, artwork_required).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Object storage is temporarily unavailable (storage_unavailable).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "IntegrationKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "IssueRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "minLength": 1
          },
          "badgeTemplateId": {
            "type": "string",
            "minLength": 1
          },
          "recipientIdentity": {
            "type": "string",
            "minLength": 1
          },
          "recipientIdentityType": {
            "type": "string",
            "enum": [
              "email",
              "email_sha256",
              "did",
              "url"
            ]
          },
          "recipientIdentifiers": {
            "maxItems": 10,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "identifierType": {
                  "type": "string",
                  "enum": [
                    "emailAddress",
                    "sourcedId",
                    "did",
                    "nationalIdentityNumber",
                    "studentId"
                  ]
                },
                "identifier": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 512
                }
              },
              "required": [
                "identifierType",
                "identifier"
              ]
            }
          },
          "recipientDisplayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "issuerImageUri": {
            "type": "string",
            "maxLength": 2048,
            "format": "uri"
          },
          "validUntil": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "idempotencyKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        },
        "required": [
          "tenantId",
          "badgeTemplateId",
          "recipientIdentity",
          "recipientIdentityType",
          "idempotencyKey"
        ],
        "additionalProperties": false
      },
      "RevokeRequest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "minLength": 1
          },
          "assertionId": {
            "type": "string",
            "minLength": 1
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          },
          "idempotencyKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        },
        "required": [
          "tenantId",
          "assertionId",
          "reason",
          "idempotencyKey"
        ]
      },
      "Accepted": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "const": "queued"
              },
              "channel": {
                "type": "string",
                "const": "programmatic_api_key"
              },
              "assertionId": {
                "type": "string",
                "minLength": 1,
                "maxLength": 256
              },
              "idempotencyKey": {
                "type": "string"
              },
              "operationId": {
                "type": "string",
                "minLength": 1,
                "maxLength": 256
              },
              "statusUrl": {
                "type": "string",
                "format": "uri"
              },
              "jobType": {
                "type": "string",
                "const": "issue_badge"
              }
            },
            "required": [
              "status",
              "channel",
              "assertionId",
              "idempotencyKey",
              "operationId",
              "statusUrl",
              "jobType"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "const": "queued"
              },
              "channel": {
                "type": "string",
                "const": "programmatic_api_key"
              },
              "assertionId": {
                "type": "string",
                "minLength": 1,
                "maxLength": 256
              },
              "idempotencyKey": {
                "type": "string"
              },
              "operationId": {
                "type": "string",
                "minLength": 1,
                "maxLength": 256
              },
              "statusUrl": {
                "type": "string",
                "format": "uri"
              },
              "jobType": {
                "type": "string",
                "const": "revoke_badge"
              },
              "revocationId": {
                "type": "string",
                "minLength": 1,
                "maxLength": 256
              }
            },
            "required": [
              "status",
              "channel",
              "assertionId",
              "idempotencyKey",
              "operationId",
              "statusUrl",
              "jobType",
              "revocationId"
            ],
            "additionalProperties": false
          }
        ]
      },
...
      "Error": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "invalid_request",
              "invalid_expiry",
              "api_key_required",
              "invalid_api_key",
              "tenant_mismatch",
              "insufficient_scope",
              "invalid_api_key_scopes",
              "api_key_owner_required",
              "idempotency_conflict",
              "template_not_found",
              "template_archived",
              "artwork_required",
              "storage_unavailable",
              "operation_not_found",
              "assertion_not_found"
            ]
          },
          "error": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "code",
          "error"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "servers": [
    {
      "url": "https://credtrail.org"
    }
  ]
}